app/requirements.txt
# Backend
click==8.1.8
dataclasses-json==0.6.7
datasets==3.6.0
docker==7.1.0
dspy==2.6.27
email-validator==2.2.0
exceptiongroup==1.3.0
Faker==25.9.2
fastapi==0.115.13
huggingface-hub==0.33.0
langgraph==0.4.8
langsmith==0.4.1
mlflow==3.1.0
mlflow-skinny==3.1.0
numpy==2.0.2
openai==1.91.0
pandas==2.3.0
pydantic==2.11.7
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.4
scikit-learn==1.6.1
scipy==1.13.1

uvicorn==0.34.3
guardrails-ai==0.6.6
# streamlit==1.46.0
psycopg2-binary